Extract output snapshot handling into dedicated module
