+++ src/main.rs
@@ -9,6 +9,7 @@ mod install;
 mod ipc;
 mod managed_network;
 mod output_capture;
+mod output_snapshot;
 mod output_stream;
 mod output_timeline;
 mod oversized_output;
